fini youpi

+++ src/Mention.java
@@ -2,9 +2,9 @@ public class Mention {
 
   private int artisteSource;
   private int artisteMentionne;
-  private int poids;
+  private double poids;
 
-  public Mention(int artisteSource, int artisteMentionne, int poids) {
+  public Mention(int artisteSource, int artisteMentionne, double poids) {
     this.artisteSource = artisteSource;
     this.artisteMentionne = artisteMentionne;
     this.poids = poids;
@@ -18,7 +18,7 @@ public class Mention {
     return artisteMentionne;
   }
 
-  public int getPoids() {
+  public double getPoids() {
     return poids;
   }
 }
